mahima/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.104.1
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.13
numpy>=2.0.0
oauthlib==3.2.2
packaging==25.0
Pillow==10.0.1
pluggy==1.6.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.3
python-docx==1.1.0
python-dotenv==1.0.0
python-linkedin-v2==0.9.4
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.27.0
thinc==8.2.5
tomli==2.2.1
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
